% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_sonar_qube.R
\name{run_lintr_check}
\alias{run_lintr_check}
\title{sonarR
this is a function named 'sonarScan' that lints your code located in the
'dir' folder and produces a JSON result file named in 'outFile'
The output will contain the linting issues of your code that can be loaded
into SonarQube}
\usage{
run_lintr_check(
  path_in = ".",
  exclusions_in = NULL,
  mapping = NULL,
  prepare_for_sonar_qube = FALSE,
  output_file_name = NULL
)
}
\description{
You can learn more about sonarR authoring:
}
\details{
-https://github.com/paulospx/sonarR

startLine and endLine always the same from line_number
startColumn always 0, endColumn length of line unless empty then 1
}
