% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lintr_check.R
\name{run_lintr_check}
\alias{run_lintr_check}
\title{lintr/sonar qube
to easily detect security vulnerabilities in your code via lintr and/or SonarQube
this is a function named 'run_lintr_check' that lints your code
produces a JSON file which contains the linting issues of your code and which can be loaded into SonarQube}
\usage{
run_lintr_check(
  path_in = ".",
  exclusions_in = NULL,
  mapping = NULL,
  prepare_for_sonar_qube = FALSE,
  output_file_name = NULL
)
}
\description{
You can learn more about sonarR authoring:
}
\details{
-https://github.com/paulospx/sonarR
}
